deduplicate shared utilities and merge identical components
Extract cleanMeaning() and pickDistractors() to shared/utils/vocab.ts,
import todayDateString from streak.ts instead of redefining in progressStore,
remove ClozeDelete (identical to FillInBlank) and route both subtypes to
FillInBlank, extract _parse_llm_json() helper in backend llm.py.
